Social Media Link specs Fix social media link specs
Spec for social media links

Fix random in community's social media services

+++ app/models/community.rb
@@ -32,6 +32,8 @@ class Community < ActiveRecord::Base
   protected
 
   def destroy_social_media_services
+    return unless social_media_services_changed?
+
     social_media_services_change.inject(:-).each do |service|
       social_media_links.where(service: service).destroy_all
     end
